+++ src/background-scripts/background.js
@@ -41,12 +41,12 @@ function onMessage(msg, sender, respond) {
 					return;
 				let TabsLeft = msg.tUrls.length;
 
-				// Reverse the url order so that we are opening in the correct order
-				for(let url of msg.tUrls.reverse()) {
+				let nbOpened = 0;
+				for(let url of msg.tUrls) {
 					let props = {
 						url:    url,
 						active: Prefs.SwitchFocusToNewTab ? (--TabsLeft) === 0 : false,	// Activate the last tab to be opened
-						index:  Prefs.OpenTabsAtEndOfTabBar ? tabsAll.length : tabs[0].index + 1, // Open tabs at the end of the tab bar
+						index:  (Prefs.OpenTabsAtEndOfTabBar ? tabsAll.length : tabs[0].index + 1) + nbOpened++, // Open tabs at the end of the tab bar
 					};
 					if(isFirefox) {
 						props.cookieStoreId = tabs[0].cookieStoreId;
@@ -69,24 +69,22 @@ function onMessage(msg, sender, respond) {
  */
 async function CheckInstallation() {
 	try {
-		let Url,
-			item     = await browser.storage.local.get('LastInstalledVersion'),
+		let item     = await browser.storage.local.get('LastInstalledVersion'),
 			manifest = browser.runtime.getManifest();
 
-		if(!item || !item.LastInstalledVersion) {
-			// New installation
-			Url = 'https://cpriest.github.io/SnapLinksPlus/#/Tutorial';
-		} else if(item.LastInstalledVersion != manifest.version) {
-			// Update/Upgrade
-			Url = 'https://cpriest.github.io/SnapLinksPlus/#/Updated';
-		}
-
-		if(Url) {
-			//noinspection ES6MissingAwait
-			browser.tabs.create({
-				url:    Url,
-				active: true,
-			});
+		if(Prefs.ShowUpdateNotification) {
+			if(!item || !item.LastInstalledVersion) {
+				// New installation
+				Notify('Snap Links Installed', `Click for a quick tutorial on usage.`, id => browser.tabs.create({
+					url:    'https://cpriest.github.io/SnapLinksPlus/#/Tutorial',
+					active: true,
+				}));
+			} else if(item.LastInstalledVersion != manifest.version) {
+				Notify('Snap Links Updated', `Version ${manifest.version} is now installed.\n\nClick here for more information.`, id => browser.tabs.create({
+					url:    'https://cpriest.github.io/SnapLinksPlus/#/Updated',
+					active: true,
+				}));
+			}
 		}
 
 		// Transition to sync storage for v3.1.7 upgrade
@@ -103,6 +101,23 @@ async function CheckInstallation() {
 	} catch(e) {
 		console.error('Error while getting LastInstalledVersion: ', e);
 	}
+}
+
+function Notify(title, message, onClick) {
+	if(onClick)
+		browser.notifications.onClicked.addListener((...args) => onClick(...args));
+
+	browser.notifications.create(
+		'', {
+			type:    'basic',
+			title:   title,
+			iconUrl: 'res/SnapLinksLogo32.png',
+			message: message,
+//			buttons:        [
+//				{ title: 'Disable This Notification' }
+//			]
+		})
+		.then((res) => { });
 }
 
 DOMReady.then(() => {